% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetBenReg.R
\name{clean_data}
\alias{clean_data}
\title{Clean Data Function}
\usage{
clean_data(data)
}
\arguments{
\item{data}{a data.frame containing data to be cleaned.}
}
\value{
a data.frame containing the cleaned data
}
\description{
This function keeps data with non-missing values only, drops categorical levels unused in final data, and drops categorical variables with only 1 level.
}
