% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetBenReg.R
\name{KM}
\alias{KM}
\title{Kaplan-Meier Function}
\usage{
KM(X, delta)
}
\arguments{
\item{X}{a vector containing continuous positive follow-up time.}

\item{delta}{a vector containing binary indicator of event, 1 - complete, 0 - censored.}
}
\value{
a vector containing the survival probability at each follow-up time
}
\description{
This function estimates the survival probability using Kaplan-Meier method, assuming censoring is not covariate-dependent. used to calculate weights of inverse probability of censoring.
}
\keyword{Kaplan-Meier}
