% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetBenReg.R
\name{NetBenReg.SW.DR}
\alias{NetBenReg.SW.DR}
\title{Doubly robust simple weighted method for net-benefit regression}
\usage{
NetBenReg.SW.DR(
  X,
  delta,
  Cost.total,
  Eff.total,
  group,
  Z = NULL,
  PS.Z = NULL,
  interaction = NULL,
  Sep.K = TRUE,
  PS.trim = 0.05,
  lambda = NULL,
  DR.Reg.SE = TRUE,
  Cost.only = FALSE
)
}
\arguments{
\item{X}{vector containing continuous positive follow-up time.}

\item{delta}{vector containing binary indicator of event (such as death), 1 - complete, 0 - censored.}

\item{Cost.total}{vector containing observed total costs.}

\item{Eff.total}{vector containing observed total effectiveness. If not provided, assume effectiveness is survival.}

\item{group}{vector containing binary treatment indicator, 1 - treatment, 0 - control.}

\item{Z}{vector, matrix, or dataframe containing covariates for net-benefit regression. If not provided, will do unadjusted analysis using simple regression. Defaults to NULL.}

\item{PS.Z}{vector, matrix, or dataframe containing covariates matrix for propensity score model using logistic regression. If not provided, will fit an unadjusted logistic regression (e.g., for randomized studies). Defaults to NULL.}

\item{interaction}{vector containing covariate names to be included in interactions with treatment, must be a subset of variable names in Z, otherwise the variable not in Z will be ignored. Defaults to NULL.}

\item{Sep.K}{logical, if TRUE, estimate K (survival function of censoring time, used in inverse probability of censoring weighting) using Kaplan-Meier estimator within each group separately. Defaults to TRUE.}

\item{PS.trim}{value between (0, 0.5) to trim extreme propensity scores outside the range of (PS.trim, 1-PS.trim). Defaults to 0.05.}

\item{lambda}{vector or scalar containing cost-effectiveness threshold values (e.g., willingness-to-pay for 1 unit of additional effectiveness).}

\item{DR.Reg.SE}{logical, if TRUE, report SE (and p-value) for the regression part of doubly robust method, otherwise only coefficient is reported. Defaults to TRUE.}

\item{Cost.only}{logical. If TRUE, fit a regression with dependent variable as cost. Defaults to FALSE.}
}
\description{
This function uses doubly robust simple weighted method for net-benefit regression.
}
\examples{
NetBenReg.SW.DR()
}
\keyword{doubly}
\keyword{net-benefit}
\keyword{regression,}
\keyword{robust}
\keyword{simple}
\keyword{weighted,}
