% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetBenReg.R
\name{NetBenReg_AL}
\alias{NetBenReg_AL}
\title{Naive method using all data ignoring censoring status for net-benefit regression}
\usage{
NetBenReg_AL(
  X,
  delta,
  Cost.total,
  Eff.total,
  group,
  Z = NULL,
  interaction = NULL,
  lambda = NULL,
  Cost.only = FALSE
)
}
\arguments{
\item{X}{a vector containing continuous positive follow-up time.}

\item{delta}{a vector containing binary indicator of event (such as death), 1 - complete, 0 - censored.}

\item{Cost.total}{a vector containing observed total costs.}

\item{Eff.total}{a vector containing observed total effectiveness. If not provided, assume effectiveness is survival.}

\item{group}{a vector containing binary treatment indicator, 1 - treatment, 0 - control.}

\item{Z}{a vector, matrix, or dataframe containing covariates for net-benefit regression. If not provided, will do unadjusted analysis using simple regression. Defaults to NULL.}

\item{interaction}{a vector containing covariate names to be included in interactions with treatment, must be a subset of variable names in Z, otherwise the variable not in Z will be ignored. Defaults to NULL.}

\item{lambda}{a vector or scalar containing cost-effectiveness threshold values (e.g., willingness-to-pay for 1 unit of additional effectiveness).}

\item{Cost.only}{logical. If TRUE, fit a regression with dependent variable as cost. Defaults to FALSE.}
}
\description{
This function uses all data ignoring censoring status for net-benefit regression, not recommended but included for comparison.
}
\keyword{all}
\keyword{censoring}
\keyword{data}
\keyword{ignoring}
\keyword{net-benefit}
\keyword{regression,}
\keyword{status}
