% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetBenReg.R
\name{NetBenReg_PT}
\alias{NetBenReg_PT}
\title{Partitioned method for net-benefit regression}
\usage{
NetBenReg_PT(
  X,
  delta,
  Cost.grp,
  Eff.grp,
  Part.times,
  group,
  Z = NULL,
  interaction = NULL,
  Sep.K = TRUE,
  lambda = NULL,
  Cost.only = FALSE,
  Aux.X = NULL,
  returnData = FALSE
)
}
\arguments{
\item{X}{a vector containing continuous positive follow-up time.}

\item{delta}{a vector containing binary indicator of event (such as death), 1 - complete, 0 - censored.}

\item{Cost.grp}{a matrix or dataframe containing observed grouped costs, Cost.grp\link{i,j} is observed cost of ith people accumulated in jth interval.}

\item{Eff.grp}{a matrix or dataframe containing observed grouped effectiveness. Eff.grp\link{i,j} is observed effectiveness of ith people accumulated in jth interval. If not provided, assume effectiveness is survival.}

\item{Part.times}{a vector containing end time points of each time interval, must be monotonically increasing.}

\item{group}{a vector containing binary treatment indicator, 1 - treatment, 0 - control.}

\item{Z}{a vector, matrix, or dataframe containing covariates for net-benefit regression. If not provided, will do unadjusted analysis using simple regression. Defaults to NULL.}

\item{interaction}{a vector containing covariate names to be included in interactions with treatment, must be a subset of variable names in Z, otherwise the variable not in Z will be ignored. Defaults to NULL.}

\item{Sep.K}{logical, if TRUE, estimate K (survival function of censoring time, used in inverse probability of censoring weighting) using Kaplan-Meier estimator within each group separately. Defaults to TRUE.}

\item{lambda}{a vector or scalar containing cost-effectiveness threshold values (e.g., willingness-to-pay for 1 unit of additional effectiveness).}

\item{Cost.only}{logical. If TRUE, fit a regression with dependent variable as cost. Defaults to FALSE.}

\item{returnData}{logical. If TRUE, return the cleaned data and some more results used in analysis. Defaults to FALSE.}

\item{Aux.X.}{a vector, matrix, or dataframe containing auxiliary covariates (e.g., for propensity model). Defaults to NULL.}
}
\description{
This function uses partitioned method for net-benefit regression.
}
\keyword{net-benefit}
\keyword{partitioned}
\keyword{regression,}
