% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetBenReg.R
\name{NetBenReg_SW_DR}
\alias{NetBenReg_SW_DR}
\title{Doubly robust simple weighted method for net-benefit regression}
\usage{
NetBenReg_SW_DR(
  X,
  delta,
  Cost.total,
  Eff.total,
  group,
  Z = NULL,
  PS.Z = NULL,
  interaction = NULL,
  Sep.K = TRUE,
  PS.trim = 0.1,
  lambda = NULL,
  Cost.only = FALSE
)
}
\arguments{
\item{X}{a vector containing continuous positive follow-up time.}

\item{delta}{a vector containing binary indicator of event (such as death), 1 - complete, 0 - censored.}

\item{Cost.total}{a vector containing observed total costs.}

\item{Eff.total}{a vector containing observed total effectiveness. If not provided, assume effectiveness is survival.}

\item{group}{a vector containing binary treatment indicator, 1 - treatment, 0 - control.}

\item{Z}{a vector, matrix, or dataframe containing covariates for net-benefit regression. If not provided, will do unadjusted analysis using simple regression. Defaults to NULL.}

\item{PS.Z}{a vector, matrix, or dataframe containing covariates matrix for propensity score model using logistic regression. If not provided, will fit an unadjusted logistic regression (e.g., for randomized studies). Defaults to NULL.}

\item{interaction}{a vector containing covariate names to be included in interactions with treatment, must be a subset of variable names in Z, otherwise the variable not in Z will be ignored. Defaults to NULL.}

\item{Sep.K}{logical, if TRUE, estimate K (survival function of censoring time, used in inverse probability of censoring weighting) using Kaplan-Meier estimator within each group separately. Defaults to TRUE.}

\item{PS.trim}{a value between (0, 0.5) to trim extreme propensity scores outside the range of (PS.trim, 1-PS.trim). Defaults to 0.1.}

\item{lambda}{a vector or scalar containing cost-effectiveness threshold values (e.g., willingness-to-pay for 1 unit of additional effectiveness).}

\item{Cost.only}{logical. If TRUE, fit a regression with dependent variable as cost. Defaults to FALSE.}
}
\description{
This function uses doubly robust simple weighted (unpartitioned version) method for net-benefit regression.
}
\keyword{doubly}
\keyword{net-benefit}
\keyword{regression,}
\keyword{robust}
\keyword{simple}
\keyword{weighted,}
