% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetBenReg.R
\name{plot.NetBenReg}
\alias{plot.NetBenReg}
\title{Create CEAC plot based on fitted net-benefit regression}
\usage{
\method{plot}{NetBenReg}(
  object,
  subgroup = NULL,
  add = FALSE,
  xlab = "Cost-effectiveness threshold",
  ylab = "Probability Treatment 1 is cost-effective",
  pch = 20,
  cex = 1,
  lty = 1,
  lwd = 1,
  type = "o",
  ylim = NULL,
  ...
)
}
\arguments{
\item{object}{a fitted "NetBenReg" model object.}

\item{subgroup}{a list of covariates to define subgroup. If there is interaction in net-benefit regression model, values for those covariates in the interaction are required.}

\item{add}{logicial. If TRUE, will add the curve to the existing plot, instead of creating a new plot. Defaults to FALSE.}

\item{xlab}{label given to the x-axis. Defaults to "Cost-effectiveness threshold".}

\item{ylab}{label given to the y-axis. Defaults to "Probability Treatment 1 is cost-effective".}

\item{pch}{a numeric value specifying the point to label the curve. The 'points' help file contains examples of the possible marks. Defaults to 20.}

\item{cex}{a numeric value specifying the size of the marks. Defaults to 1.}

\item{lty}{a vector of integers specifying line types for each curve. Defaults to 1.}

\item{lwd}{a vector of numeric values for line widths. Defaults to 1.}

\item{type}{character indicating the type of plotting. The 'plot.default' help file contains more explanatuib. Defaults to 'o'.}

\item{...}{other arguments that will be passed forward to the underlying plot.default method.}
}
\description{
Creating cost-effectiveness acceptability curve (CEAC) plot based on the fitted net-benefit regression.
}
\examples{
data(CEdata)
lambda=seq(0,6,0.5)

# for non-DR method without interaction or DR method, one curve for overall
fit1 <- NetBenReg(Followup=CEdata$survival, delta=CEdata$dead, group=CEdata$Trt, Cost=CEdata[,8:22], Eff=CEdata[,24:38],
	Part.times=1:15, Method='PT', Z=CEdata[,5:7], interaction=c("LBBB"), PS.Z=CEdata[,5:7],	Doubly.Robust=TRUE,
	Eff.only=TRUE, lambda=lambda, L=10)
plot(fit1,ylab="Probability new treatment is cost-effective",xlab="Cost-effectiveness threshold (in $1000) for one additional QALY",lwd=2, pch=19,cex=1.2)

# for non-DR method with interaction, one curve for each subgroup defined by interactions
fit2 <- NetBenReg(Followup=CEdata$survival, delta=CEdata$dead, group=CEdata$Trt, Cost=CEdata[,8:22], Eff=CEdata[,24:38], 
	Part.times=1:15, Method='PT', Z=CEdata[,5:7], interaction=c("LBBB","female"), Eff.only=TRUE, lambda=lambda, L=10)
plot(fit2,subgroup=list(LBBB=0,female=0),add=TRUE,col="gray50",lwd=2,lty=2,pch=15,cex=1.2)		#subgroup of non-LBBB male, adjusted for age
plot(fit2,subgroup=list(LBBB=1,female=1),add=TRUE,col="gray50",lwd=2,lty=3,pch=17,cex=1.2)		#subgroup of LBBB female, adjusted for age 
}
\keyword{acceptability}
\keyword{cost-effectiveness}
\keyword{curve,}
\keyword{net-benefit}
\keyword{regression}
