% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetBenReg.R
\name{plot.NetBenRegDR}
\alias{plot.NetBenRegDR}
\title{Create CEAC plot based on fitted net-benefit regression with doubly robust method}
\usage{
\method{plot}{NetBenRegDR}(
  object,
  add = FALSE,
  xlab = "WTP",
  ylab = "Probability Treatment 1 is cost-effective",
  pch = 20,
  cex = 1,
  lty = 1,
  lwd = 1,
  type = "o",
  ...
)
}
\arguments{
\item{object}{a fitted "NetBenRegDR" model object.}

\item{add}{logicial. If TRUE, will add the curve to the existing plot, instead of creating a new plot. Defaults to FALSE.}

\item{...}{other graphical parameters for plot().}
}
\description{
This function creates cost-effectiveness acceptability curve (CEAC) plot based on fitted net-benefit regression with doubly robust method.
}
\examples{
plot.NetBenRegDR()
}
\keyword{acceptability}
\keyword{cost-effectiveness}
\keyword{curve,}
\keyword{doubly}
\keyword{net-benefit}
\keyword{regression,}
\keyword{robust}
