% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetBenReg.R
\name{print.NetBenReg}
\alias{print.NetBenReg}
\title{Print the summary results of the fitted net-benefit regression}
\usage{
\method{print}{NetBenReg}(object)
}
\arguments{
\item{object}{a fitted "NetBenReg" model object.}
}
\description{
Printing the summary results of the fitted net-benefit regression.
}
\examples{
data(CEdata)	
fit1 <- NetBenReg(Followup=CEdata$survival, delta=CEdata$dead, group=CEdata$Trt, Cost=CEdata[,8:22], Eff=CEdata[,24:38], 	
	Part.times=1:15, Method='SW', Z=CEdata[,5:7], Eff.only=TRUE, Cost.only=TRUE, lambda=seq(0,6,3), L=10)
print(fit1)
}
\keyword{net-benefit}
\keyword{regression}
\keyword{summary,}
