% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetBenReg.R
\name{KM}
\alias{KM}
\title{Kaplan-Meier Function}
\usage{
KM(X, delta)
}
\arguments{
\item{X}{vector containing continuous positive follow-up time.}

\item{delta}{vector containing binary indicator of event, 1 - complete, 0 - censored.}
}
\description{
This function estimates the survival probability using Kaplan-Meier method, used to calculate weights of inverse probability of censoring.
}
\examples{
KM()
}
\keyword{Kaplan-Meier}
