% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetBenReg.R
\name{NetBenReg}
\alias{NetBenReg}
\title{Net-benefit regression with possibly censored cost-effectiveness data}
\usage{
NetBenReg(
  Followup,
  delta,
  group,
  Cost = NULL,
  Eff = NULL,
  Patition.times = NULL,
  Z = NULL,
  PS.Z = NULL,
  interaction = NULL,
  Method = c("SW", "PT", "CC", "AL"),
  Sep.K = TRUE,
  PS.trim = 0.05,
  Doubly.Robust = FALSE,
  DR.Reg.SE = TRUE,
  Eff.only = FALSE,
  Cost.only = FALSE,
  lambda = NULL,
  L
)
}
\arguments{
\item{Followup}{vector containing continuous positive follow-up time of length n.}

\item{delta}{vector containing binary indicator of event (such as death) of length n, 1 - complete, 0 - censored.}

\item{group}{vector containing binary treatment indicator, 1 - considered treatment, 0 - comparison.}

\item{Cost}{vector, n x m matrix or dataframe containing observed total or grouped costs, Cost\link{i,j} is observed cost of the ith people accumulated in the jth interval. Can be a vector or a one-column matrix for total costs only.}

\item{Eff}{vector, n x m matrix or dataframe containing observed total or grouped costs, Eff\link{i,j} is observed effectiveness of the ith people accumulated in the jth interval. Can be a vector or a one-column matrix for total effectiveness only. If not provided, assume effectiveness is survival.}

\item{Patition.times}{vector of length m, containing end time points (in days) of each time interval. Must be monotonically increasing (eg, 1st interval is Day 1 to Patition.times\link{1}, 2nd is interval Day Patition.times\link{1}+1 to Patition.times\link{2}). Required if using PT method without Eff provided, or truncating grouped costs and effectiveness if they may be partially outside time limit L.}

\item{Z}{vector, n x p matrix or dataframe containing covariates for net-benefit regression. If not provided, will do unadjusted analysis using simple regression. Defaults to NULL.}

\item{PS.Z}{vector, matrix, or dataframe containing covariates matrix for propensity score model using logistic regression. If not provided, will fit an unadjusted logistic regression (e.g., for randomized studies). Defaults to NULL.}

\item{interaction}{vector containing covariate names to be included in interactions with treatment, must be a subset of variable names in Z, otherwise the variable not in Z will be ignored. Defaults to NULL.}

\item{Method}{method for estimation. 'SW' - simple weighted, 'PT' - partitioned, 'CC' - naive complete case only, 'AL' - naive all data ignoring censoring status, doubly robust method requires either 'SW' or 'PT'.}

\item{Sep.K}{logical, if TRUE, estimate K (survival function of censoring time, used in inverse probability of censoring weighting) using Kaplan-Meier estimator within each group separately. Defaults to TRUE.}

\item{PS.trim}{value between (0, 0.5) to trim extreme propensity scores outside the range of (PS.trim, 1-PS.trim). Defaults to 0.05.}

\item{Doubly.Robust}{logical, if TRUE, perform doubly robust method. Defaults to FALSE.}

\item{DR.Reg.SE}{logical, if TRUE, report SE (and p-value) for the regression part of doubly robust method, otherwise only coefficient is reported. Defaults to TRUE.}

\item{Eff.only}{logical. If TRUE, fit a regression with dependent variable as effect. Defaults to FALSE.}

\item{Cost.only}{logical. If TRUE, fit a regression with dependent variable as cost. Defaults to FALSE.}

\item{lambda}{vector or scalar containing cost-effectiveness threshold values (e.g., willingness-to-pay for 1 unit of additional effectiveness).}

\item{L}{time limit horizon, used to truncate event time (equivalent to having an event at L if the patient is alive at L), costs and effectiveness if they are outside this time limit, assuming cost and effectiveness are evenly spread within each time interval during truncation.}
}
\description{
This function fits net-benefit regression with possibly censored cost-effectiveness data.
}
\examples{
NetBenReg()
}
\keyword{censoring}
\keyword{net-benefit}
\keyword{regression,}
