% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetBenReg.R
\name{plot.NetBenReg}
\alias{plot.NetBenReg}
\title{Create CEAC plot based on fitted net-benefit regression}
\usage{
\method{plot}{NetBenReg}(
  object,
  subgroup = NULL,
  add = FALSE,
  xlab = "WTP",
  ylab = "Probability Treatment 1 is cost-effective",
  pch = 20,
  cex = 1,
  lty = 1,
  lwd = 1,
  type = "o",
  ...
)
}
\arguments{
\item{object}{a fitted "NetBenReg" model object.}

\item{subgroup}{a list of covariates to define subgroup. If there is interaction in net-benefit regression model, values for those covariates in the interaction are required.}

\item{add}{logicial. If TRUE, will add the curve to the existing plot, instead of creating a new plot. Defaults to FALSE.}

\item{...}{other graphical parameters for plot().}
}
\description{
This function creates cost-effectiveness acceptability curve (CEAC) plot based on fitted net-benefit regression (without doubly robust method).
}
\examples{
plot.NetBenReg()
}
\keyword{acceptability}
\keyword{cost-effectiveness}
\keyword{curve,}
\keyword{net-benefit}
\keyword{regression}
